refactor: Info class is moved to info.hpp
